refactor: separate main function to its own file
